Updated Searching
Improved query_engine.py so that it no longer corrects words if they're not in the dictionary so long as they are indexed in the database. Also added various improvements to index.html which now has a history drop down menu, functional forward and back buttons, searching from the URL, and more.